PostalItemController

+++ src/main/java/com/github/ajourer/postalservice/PostOfficeRepository.java
@@ -2,6 +2,6 @@ package com.github.ajourer.postalservice;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostOfficeRepository extends CrudRepository<PostOffice, Long> {
-    public PostOffice findById(long id);
+public interface PostOfficeRepository extends CrudRepository<PostOffice, Integer> {
+    public PostOffice findById(int id);
 }
